refactor(webapp): address PR review on usage breakdown (EXT-1144)
- Break ChartCard into focused pieces: useChartInteractions (legend/hover
  state), useChartData (per-day rows), plus BreakdownChart and ChartLegend
  components; ChartCard is now a thin composition.
- Extract the panel header controls into a BreakdownControls component.
- Fold the Status semantic colors into colorForValue; note the interim hex
  palette for promotion to design-system tokens (NAN-5927).
- Drop ChartCard's totalOverride prop (fixtures discrepancy accepted).
- Clear the hovered series when one is hidden, so the chart no longer stays
  dimmed by a now-hidden hover target.
